- Updating randomuser api url to the fixed version of 1.1 - Moving a test class to the correct package
